refactor: update HistoryManager to use shared_ptr for better memory management and modify related methods, add Builder

+++ src/region/ConvexRegion.cpp
@@ -276,12 +276,12 @@ void ConvexRegion::forEachLine(
     }
 }
 
-uint64 ConvexRegion::size() const {
+size_t ConvexRegion::size() const {
     double volume = 0;
     for (auto triangle : triangles) {
         volume +=
             triangle.getVertex(0).cross(triangle.getVertex(1)).dot(triangle.getVertex(2));
     }
-    return (uint64)std::abs(volume / 6.0);
+    return (size_t)std::abs(volume / 6.0);
 };
 } // namespace we
