+++ src/region/PolyRegion.h
@@ -26,12 +26,12 @@ public:
 
     RegionType getType() const override { return RegionType::Poly; }
 
-    uint64 size() const override;
+    size_t size() const override;
 
     bool removePoint(std::optional<BlockPos> const&) override;
 
-    void
-    forEachBlockUVInRegion(std::function<void(BlockPos const&, double, double)>&& todo
+    void forEachBlockUVInRegion(
+        std::function<void(BlockPos const&, double, double)>&& todo
     ) const override;
 
     Vec3 getCenter() const override;
